# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EduJobs Connect API",
    "description": "Comprehensive API for the EduJobs Connect platform - connecting Rwandan students with universities and job opportunities",
    "version": "1.0.0",
    "contact": {
      "name": "EduJobs Connect Support",
      "email": "support@edujobsconnect.rw",
      "url": "https://edujobsconnect.rw"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.edujobsconnect.rw/api",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Enter your token in the text input below."
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "objectId",
            "example": "507f1f77bcf86cd799439011"
          },
          "firstName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "example": "Doe"
          },
          "fullName": {
            "type": "string",
            "readOnly": true,
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "john.doe@example.com"
          },
          "role": {
            "type": "string",
            "enum": ["student", "employer", "admin"],
            "default": "student",
            "example": "student"
          },
          "status": {
            "type": "string",
            "enum": ["active", "inactive", "suspended"],
            "default": "active",
            "example": "active"
          },
          "isEmailVerified": {
            "type": "boolean",
            "example": true
          },
          "lastLogin": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "profile": {
            "$ref": "#/components/schemas/UserProfile"
          },
          "preferences": {
            "$ref": "#/components/schemas/UserPreferences"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "required": ["firstName", "lastName", "email"]
      },
      "UserProfile": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+250788123456"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "example": "1995-06-15"
          },
          "gender": {
            "type": "string",
            "enum": ["male", "female", "other"],
            "example": "male"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "bio": {
            "type": "string",
            "maxLength": 500,
            "example": "Computer Science student passionate about technology and innovation."
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["JavaScript", "React", "Node.js"]
          },
          "experience": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Experience"
            }
          },
          "education": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Education"
            }
          },
          "resume": {
            "type": "string",
            "format": "uri",
            "example": "/uploads/resumes/john_doe_resume.pdf"
          },
          "avatar": {
            "type": "string",
            "format": "uri",
            "example": "/uploads/avatars/john_doe.jpg"
          }
        }
      },
      "UserPreferences": {
        "type": "object",
        "properties": {
          "emailNotifications": {
            "type": "boolean",
            "default": true,
            "example": true
          },
          "smsNotifications": {
            "type": "boolean",
            "default": false,
            "example": false
          },
          "jobAlerts": {
            "type": "boolean",
            "default": true,
            "example": true
          },
          "universityAlerts": {
            "type": "boolean",
            "default": true,
            "example": true
          },
          "language": {
            "type": "string",
            "default": "en",
            "example": "en"
          },
          "timezone": {
            "type": "string",
            "default": "Africa/Kigali",
            "example": "Africa/Kigali"
          }
        }
      },
      "Address": {
        "type": "object",
        "properties": {
          "street": {
            "type": "string",
            "example": "123 Main Street"
          },
          "city": {
            "type": "string",
            "example": "Kigali"
          },
          "province": {
            "type": "string",
            "example": "Kigali City"
          },
          "country": {
            "type": "string",
            "default": "Rwanda",
            "example": "Rwanda"
          },
          "postalCode": {
            "type": "string",
            "example": "00000"
          }
        }
      },
      "Experience": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Software Developer Intern"
          },
          "company": {
            "type": "string",
            "example": "Tech Solutions Ltd"
          },
          "location": {
            "type": "string",
            "example": "Kigali, Rwanda"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "example": "2023-06-01"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "example": "2023-08-31"
          },
          "current": {
            "type": "boolean",
            "default": false,
            "example": false
          },
          "description": {
            "type": "string",
            "example": "Developed web applications using React and Node.js"
          }
        }
      },
      "Education": {
        "type": "object",
        "properties": {
          "institution": {
            "type": "string",
            "example": "University of Rwanda"
          },
          "degree": {
            "type": "string",
            "example": "Bachelor of Computer Science"
          },
          "fieldOfStudy": {
            "type": "string",
            "example": "Computer Science"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "example": "2020-09-01"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "example": "2024-06-30"
          },
          "current": {
            "type": "boolean",
            "default": false,
            "example": false
          },
          "gpa": {
            "type": "string",
            "example": "3.8"
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "example": "password123"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["firstName", "lastName", "email", "password"],
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)",
            "example": "Password123"
          },
          "role": {
            "type": "string",
            "enum": ["student", "employer"],
            "default": "student",
            "example": "student"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Login successful"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Validation failed"
          },
          "message": {
            "type": "string",
            "example": "Please check your input data"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user",
        "description": "Create a new user account with email verification",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or user already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login user",
        "description": "Authenticate user and return access tokens",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or account locked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "423": {
            "description": "Account locked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Refresh access token",
        "description": "Generate new access token using refresh token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Token refreshed successfully"
                    },
                    "token": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Send password reset email",
        "description": "Send password reset link to user's email",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "john.doe@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reset email sent (if account exists)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "If an account with that email exists, a password reset link has been sent"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Reset password with token",
        "description": "Reset user password using reset token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token", "password"],
                "properties": {
                  "token": {
                    "type": "string",
                    "example": "reset_token_here"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6,
                    "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)",
                    "example": "NewPassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Password has been reset successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify-email": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Verify email address",
        "description": "Verify user email using verification token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token"],
                "properties": {
                  "token": {
                    "type": "string",
                    "example": "email_verification_token"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Email verified successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get current user profile",
        "description": "Get detailed information about the authenticated user",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Logout user",
        "description": "Logout user (client-side token removal)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logged out successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Logged out successfully"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}